+++ src/index.js
@@ -2,21 +2,11 @@
 // @ts-check
 
 const debug = require('debug')('cypress-split')
-const { getSpecs } = require('find-cypress-specs')
 const ghCore = require('@actions/core')
-const { getChunk } = require('./chunk')
-const { findFile } = require('./find-file')
+
 const { parseSplitInputs, getSpecsToSplit } = require('./parse-inputs')
-const {
-  splitByDuration,
-  hasTimeDifferences,
-  mergeTimings,
-} = require('./timings')
-const {
-  getEnvironmentFlag,
-  printSpecsList,
-  printSpecsListWithDurations,
-} = require('./utils')
+const { hasTimeDifferences, mergeTimings } = require('./timings')
+const { getEnvironmentFlag, splitSpecsLogic } = require('./utils')
 const path = require('path')
 const os = require('os')
 const fs = require('fs')
@@ -108,11 +98,6 @@ function cypressSplit(on, config) {
   }
 
   if (isDefined(SPLIT) && isDefined(SPLIT_INDEX)) {
-    // if (!specs) {
-    //   const returnAbsolute = true
-    //   // @ts-ignore
-    //   specs = getSpecs(config, undefined, returnAbsolute)
-    // }
     const specs = getSpecsToSplit(process.env, config)
 
     console.log('%s there are %d found specs', label, specs.length)
@@ -122,79 +107,87 @@ function cypressSplit(on, config) {
     console.log('%s chunk %d of %d', label, splitIndex + 1, splitN)
 
     debug('get chunk %o', { specs, splitN, splitIndex })
-    /** @type {string[]} absolute spec filenames */
-    let splitSpecs
+    /* @type {string[]} absolute spec filenames */
+    // let splitSpecs
 
-    const cwd = process.cwd()
-    console.log('%s specs from the current directory %s', label, cwd)
+    // const cwd = process.cwd()
+    // console.log('%s specs from the current directory %s', label, cwd)
 
-    let foundSplitFile
-    if (SPLIT_FILE) {
-      debug('loading split file %s', SPLIT_FILE)
-      try {
-        foundSplitFile = findFile(SPLIT_FILE)
-        if (!foundSplitFile) {
-          throw new Error(
-            `Could not find ${SPLIT_FILE} based on the current working directory ${cwd}`,
-          )
-        }
-        const splitFile = JSON.parse(fs.readFileSync(foundSplitFile, 'utf8'))
-        const previousDurations = splitFile.durations
-        const averageDuration =
-          previousDurations
-            .map((item) => item.duration)
-            .reduce((sum, duration) => (sum += duration), 0) /
-          previousDurations.length
-        const specsWithDurations = specs.map((specName) => {
-          const relativeSpec = path.relative(cwd, specName)
-          const foundInfo = previousDurations.find(
-            (item) => item.spec === relativeSpec,
-          )
-          if (!foundInfo) {
-            return {
-              specName,
-              duration: Math.round(averageDuration),
-            }
-          } else {
-            return {
-              specName,
-              duration: Math.round(foundInfo.duration),
-            }
-          }
-        })
-        debug('splitting by duration %d ways', splitN)
-        debug(specsWithDurations)
-        const { chunks, sums } = splitByDuration(splitN, specsWithDurations)
-        debug('split by duration')
-        debug(chunks)
-        debug('sums of durations for chunks')
-        debug(sums)
+    const { splitSpecs, foundSplitFile } = splitSpecsLogic({
+      specs,
+      splitN,
+      splitIndex,
+      splitFileName: SPLIT_FILE,
+      label,
+    })
 
-        splitSpecs = chunks[splitIndex].map((item) => item.specName)
-        console.log(
-          '%s split %d specs using durations from %s file',
-          label,
-          specsWithDurations.length,
-          SPLIT_FILE,
-        )
-        console.log(
-          '%s approximate total duration for current chunk is %s (plus Cypress overhead)',
-          label,
-          humanizeDuration(Math.round(sums[splitIndex])),
-        )
+    // let foundSplitFile
+    // if (SPLIT_FILE) {
+    //   debug('loading split file %s', SPLIT_FILE)
+    //   try {
+    //     foundSplitFile = findFile(SPLIT_FILE)
+    //     if (!foundSplitFile) {
+    //       throw new Error(
+    //         `Could not find ${SPLIT_FILE} based on the current working directory ${cwd}`,
+    //       )
+    //     }
+    //     const splitFile = JSON.parse(fs.readFileSync(foundSplitFile, 'utf8'))
+    //     const previousDurations = splitFile.durations
+    //     const averageDuration =
+    //       previousDurations
+    //         .map((item) => item.duration)
+    //         .reduce((sum, duration) => (sum += duration), 0) /
+    //       previousDurations.length
+    //     const specsWithDurations = specs.map((specName) => {
+    //       const relativeSpec = path.relative(cwd, specName)
+    //       const foundInfo = previousDurations.find(
+    //         (item) => item.spec === relativeSpec,
+    //       )
+    //       if (!foundInfo) {
+    //         return {
+    //           specName,
+    //           duration: Math.round(averageDuration),
+    //         }
+    //       } else {
+    //         return {
+    //           specName,
+    //           duration: Math.round(foundInfo.duration),
+    //         }
+    //       }
+    //     })
+    //     debug('splitting by duration %d ways', splitN)
+    //     debug(specsWithDurations)
+    //     const { chunks, sums } = splitByDuration(splitN, specsWithDurations)
+    //     debug('split by duration')
+    //     debug(chunks)
+    //     debug('sums of durations for chunks')
+    //     debug(sums)
 
-        printSpecsListWithDurations(chunks[splitIndex])
-      } catch (err) {
-        console.error('%s Could not split specs by duration', label)
-        console.error(err.message)
-        console.error('%s splitting as is by name', label)
-        splitSpecs = getChunk(specs, splitN, splitIndex)
-        printSpecsList(splitSpecs)
-      }
-    } else {
-      splitSpecs = getChunk(specs, splitN, splitIndex)
-      printSpecsList(splitSpecs)
-    }
+    //     splitSpecs = chunks[splitIndex].map((item) => item.specName)
+    //     console.log(
+    //       '%s split %d specs using durations from %s file',
+    //       label,
+    //       specsWithDurations.length,
+    //       SPLIT_FILE,
+    //     )
+    //     console.log(
+    //       '%s approximate total duration for current chunk is %s (plus Cypress overhead)',
+    //       label,
+    //       humanizeDuration(Math.round(sums[splitIndex])),
+    //     )
+
+    //     printSpecsListWithDurations(chunks[splitIndex])
+    //   } catch (err) {
+    //     console.error('%s Could not split specs by duration', label)
+    //     console.error(err.message)
+    //     console.error('%s splitting as is by name', label)
+    //     splitSpecs = getChunk(specs, splitN, splitIndex)
+    //     printSpecsList(splitSpecs)
+    //   }
+    // } else {
+    //   splitSpecs = getChunk(specs, splitN, splitIndex)
+    //   printSpecsList(splitSpecs)
+    // }
     debug('split specs')
     debug(splitSpecs)
